backend/requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.2.0
pytesseract==0.2.7
Werkzeug==0.15.5